Introducing Octothorpe.
I ported the program to C# (and refactored code) as an experiment.
Octothorpe.Core contains all of the magic; the other projects in the
solution are interfaces.

Although stable, this is still something I'm experimenting with. The
only thing I couldn't quite figure out was the model field's behavior
on Mac OS X.