Experimenting
New branch for experimenting! Things may or may not make it to a future release.

I tried to make the parser a little more efficient with loading the PLIST once, unless the source is changed. This will cut down on the file and network I/O requests, but it uses more memory to remember the assets.