Added support for getting delta updates from Pallas.
Both the "ProductVersion" and "BuildVersion" parameters are required for getting delta updates with a Pallas request.

For now, the OS version and build numbers must be manually entered and match up correctly. (i.e. Don't mix "13.5" and "99Z123" as it won't work.) The plan is to add missing version info to BuildInfo.plist and eventually, given a device/model combination, have the parser probe version/build numbers to get each delta.